Add theme-based text color and fix spin
Disclaimer: This is bad practice!

HOWEVER! I noticed that the spinning animation was sorta off. It took me a while to figure it out, but eventually I noticed that the spinning icon provided by FontAwesome had the dimensions 28x29.

Since width and height wasn't the same it create a sort of jumping (?) and made the animation less smooth.

To fix it I checked what font-size it currently had and literally tried some numbers above and below it.

My first attempt was `2.072`, which made the icon 29x29 - which didn't solve the issue, since the animation wasn't perfectly centered anymore.

`1.99999999`, as unfortunate it is, turned the icon into 28x28 and fixed the animation.