Input focus (#146)
* add focusInput functions

`autofocus` on inputs doesn't seem to work. So I manually set `focus()` whenever `open()`/`clear()`/`reset()` is called.

* set focus to searchInput after adding a game (WIP)

I know too little about Vue.js, but I'm certain that it's a bad practice to look 2 parents up to find the `searchInput` ref. Either the ref needs to be added to its children or `focusInput()` needs to be added to `GameCard.js` to use as a mixin.

* linting

* remove unnecessary if statements

* remove comments